Refactor: Verification & Nft Directory
