+++ green/src/main/java/com/greengrim/green/core/hiding/nft/NftHiding.java
@@ -1,14 +1,10 @@
 package com.greengrim.green.core.hiding.nft;
 
 import com.greengrim.green.common.entity.BaseTime;
-import com.greengrim.green.core.member.Member;
-import com.greengrim.green.core.nft.Nft;
 import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.ManyToOne;
 import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
