+++ green/src/main/java/com/greengrim/green/core/nft/controller/GetNftController.java
@@ -4,7 +4,7 @@ import com.greengrim.green.common.entity.NftSortOption;
 import com.greengrim.green.common.oauth.auth.CurrentMember;
 import com.greengrim.green.common.entity.dto.PageResponseDto;
 import com.greengrim.green.core.member.Member;
-import com.greengrim.green.core.nft.NftGrade;
+import com.greengrim.green.core.nft.entity.NftGrade;
 import com.greengrim.green.core.nft.dto.NftResponseDto.HotNftInfo;
 import com.greengrim.green.core.nft.dto.NftResponseDto.NftAndMemberInfo;
 import com.greengrim.green.core.nft.dto.NftResponseDto.NftCollectionInfo;
