+++ green/src/main/java/com/greengrim/green/core/nft/entity/NftGrade.java
@@ -1,4 +1,4 @@
-package com.greengrim.green.core.nft;
+package com.greengrim.green.core.nft.entity;
 
 public enum NftGrade {
     BASIC, STANDARD, PREMIUM
