+++ green/src/main/java/com/greengrim/green/core/nft/service/GetNftService.java
@@ -8,8 +8,8 @@ import com.greengrim.green.common.entity.dto.PageResponseDto;
 import com.greengrim.green.common.exception.BaseException;
 import com.greengrim.green.common.exception.errorCode.NftErrorCode;
 import com.greengrim.green.core.member.Member;
-import com.greengrim.green.core.nft.Nft;
-import com.greengrim.green.core.nft.NftGrade;
+import com.greengrim.green.core.nft.entity.Nft;
+import com.greengrim.green.core.nft.entity.NftGrade;
 import com.greengrim.green.core.nft.dto.NftResponseDto.HotNftInfo;
 import com.greengrim.green.core.nft.dto.NftResponseDto.NftAndMemberInfo;
 import com.greengrim.green.core.nft.dto.NftResponseDto.NftCollectionInfo;
@@ -18,7 +18,7 @@ import com.greengrim.green.core.nft.dto.NftResponseDto.NftStockAmountInfo;
 import com.greengrim.green.core.nft.dto.NftResponseDto.NftStockInfo;
 import com.greengrim.green.core.nft.repository.NftRepository;
 import com.greengrim.green.core.nft.usecase.GetNftUseCase;
-import com.greengrim.green.core.nftlike.LikeService;
+import com.greengrim.green.core.nftlike.service.LikeService;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
