+++ green/src/main/java/com/greengrim/green/core/nft/service/RegisterNftService.java
@@ -6,8 +6,8 @@ import com.greengrim.green.common.fcm.FcmService;
 import com.greengrim.green.common.web3j.Abi;
 import com.greengrim.green.core.member.Member;
 import com.greengrim.green.core.member.repository.MemberRepository;
-import com.greengrim.green.core.nft.Nft;
-import com.greengrim.green.core.nft.NftGrade;
+import com.greengrim.green.core.nft.entity.Nft;
+import com.greengrim.green.core.nft.entity.NftGrade;
 import com.greengrim.green.core.nft.repository.NftRepository;
 import com.greengrim.green.core.nft.usecase.RegisterNftUseCase;
 import jakarta.transaction.Transactional;
