+++ green/src/main/java/com/greengrim/green/core/nft/service/UpdateNftService.java
@@ -3,12 +3,12 @@ package com.greengrim.green.core.nft.service;
 import com.greengrim.green.common.exception.BaseException;
 import com.greengrim.green.common.exception.errorCode.NftErrorCode;
 import com.greengrim.green.core.member.Member;
-import com.greengrim.green.core.nft.Nft;
+import com.greengrim.green.core.nft.entity.Nft;
 import com.greengrim.green.core.nft.dto.NftRequestDto.NftModifyInfo;
 import com.greengrim.green.core.nft.dto.NftResponseDto.NftDetailInfo;
 import com.greengrim.green.core.nft.repository.NftRepository;
 import com.greengrim.green.core.nft.usecase.UpdateNftUseCase;
-import com.greengrim.green.core.nftlike.LikeService;
+import com.greengrim.green.core.nftlike.service.LikeService;
 import jakarta.transaction.Transactional;
 import java.util.List;
 import java.util.Objects;
