+++ green/src/main/java/com/greengrim/green/core/nft/usecase/GetNftUseCase.java
@@ -3,8 +3,8 @@ package com.greengrim.green.core.nft.usecase;
 import com.greengrim.green.common.entity.NftSortOption;
 import com.greengrim.green.common.entity.dto.PageResponseDto;
 import com.greengrim.green.core.member.Member;
-import com.greengrim.green.core.nft.Nft;
-import com.greengrim.green.core.nft.NftGrade;
+import com.greengrim.green.core.nft.entity.Nft;
+import com.greengrim.green.core.nft.entity.NftGrade;
 import com.greengrim.green.core.nft.dto.NftResponseDto.HotNftInfo;
 import com.greengrim.green.core.nft.dto.NftResponseDto.NftAndMemberInfo;
 import com.greengrim.green.core.nft.dto.NftResponseDto.NftCollectionInfo;
