+++ green/src/main/java/com/greengrim/green/core/nft/usecase/RegisterNftUseCase.java
@@ -1,7 +1,7 @@
 package com.greengrim.green.core.nft.usecase;
 
 import com.greengrim.green.core.member.Member;
-import com.greengrim.green.core.nft.Nft;
+import com.greengrim.green.core.nft.entity.Nft;
 
 public interface RegisterNftUseCase {
