+++ green/src/main/java/com/greengrim/green/core/nftlike/repository/LikeRepository.java
@@ -1,7 +1,9 @@
-package com.greengrim.green.core.nftlike;
+package com.greengrim.green.core.nftlike.repository;
 
-import com.greengrim.green.core.nft.Nft;
+import com.greengrim.green.core.nft.entity.Nft;
 import java.util.Optional;
+
+import com.greengrim.green.core.nftlike.entity.Like;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface LikeRepository extends JpaRepository<Like, Long> {
