+++ green/src/main/java/com/greengrim/green/core/verification/controller/VerificationController.java
@@ -1,7 +1,8 @@
-package com.greengrim.green.core.verification;
+package com.greengrim.green.core.verification.controller;
 
 import com.greengrim.green.common.oauth.auth.CurrentMember;
 import com.greengrim.green.core.member.Member;
+import com.greengrim.green.core.verification.service.VerificationService;
 import com.greengrim.green.core.verification.dto.VerificationRequestDto;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
