+++ green/src/main/java/com/greengrim/green/core/verification/repository/VerificationRepository.java
@@ -1,6 +1,6 @@
-package com.greengrim.green.core.verification;
+package com.greengrim.green.core.verification.repository;
 
-import com.greengrim.green.core.verification.Verification;
+import com.greengrim.green.core.verification.entity.Verification;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.query.Param;
